+++ src/py42/services/storage/exfiltrateddata.py
@@ -1,5 +1,3 @@
-from urllib.parse import quote
-
 from py42.services import BaseService
 
 
@@ -7,31 +5,21 @@ class ExfiltratedDataService(BaseService):
 
     _base_uri = "api/v1/"
 
-    def __init__(self, main_session, streaming_session):
-        super().__init__(main_session)
+    def __init__(self, main_connection, streaming_session):
+        super().__init__(main_connection)
         self._streaming_session = streaming_session
 
-    def get_download_token(
-        self, event_id, device_id, file_path, file_sha256, timestamp
-    ):
+    def get_download_token(self, downloadRequestUrl):
         """Get EDS download token for a file.
 
         Args:
-            event_id (str): Id of the file event that references the file desired for download.
-            device_id (str): Id of the device on which the file desired for download is stored.
-            file_path (str): Path where the file desired for download resides on the device.
-            timestamp (int): Last updated timestamp of the file in milliseconds.
+            downloadRequestUrl (str): The download request url to get the token
 
         Returns:
             :class:`py42.response.Py42Response`: A response containing download token for the file.
         """
-        params = "deviceUid={}&eventId={}&filePath={}&fileSHA256={}&versionTimestamp={}"
-        params = params.format(
-            device_id, event_id, quote(file_path), file_sha256, timestamp
-        )
-        resource = "file-download-token"
         headers = {"Accept": "*/*"}
-        uri = f"{self._base_uri}{resource}?{params}"
+        uri = f"{downloadRequestUrl}"
         return self._connection.get(uri, headers=headers)
 
     def get_file(self, token):
@@ -43,10 +31,6 @@ class ExfiltratedDataService(BaseService):
         Returns:
             Returns a stream of the file indicated by the input token.
         """
-        resource = "get-file"
-        uri = f"{self._connection.host_address}/{self._base_uri}{resource}"
-        params = {"token": token}
+        uri = f"{token}"
         headers = {"Accept": "*/*"}
-        return self._streaming_session.get(
-            uri, params=params, headers=headers, stream=True
-        )
+        return self._streaming_session.get(uri, headers=headers, stream=True)