+++ tests/services/storage/test_exfiltrateddata.py
@@ -11,21 +11,6 @@ class TestExfiltratedDataService:
         request.return_value = b"stream"
         return request
 
-    def test_get_download_token_calls_get_with_valid_params(
-        self, mock_successful_connection
-    ):
-        service = ExfiltratedDataService(
-            mock_successful_connection, mock_successful_connection
-        )
-        service.get_download_token(
-            "testeventid", "testdeviceid", "testfilepath", "testSHA256", 1223
-        )
-        qry = "deviceUid=testdeviceid&eventId=testeventid&filePath=testfilepath&fileSHA256=testSHA256&versionTimestamp=1223"
-        expected = f"api/v1/file-download-token?{qry}"
-        mock_successful_connection.get.assert_called_once_with(
-            expected, headers={"Accept": "*/*"}
-        )
-
     def test_get_file_calls_get_with_valid_params(
         self, mock_successful_connection, mock_request
     ):
@@ -33,10 +18,9 @@ class TestExfiltratedDataService:
         service = ExfiltratedDataService(
             mock_successful_connection, mock_successful_connection
         )
-        service.get_file("testtoken")
+        service.get_file("https://example.com/testpath?token=testtoken")
         mock_successful_connection.get.assert_called_once_with(
-            "https://example.com/api/v1/get-file",
+            "https://example.com/testpath?token=testtoken",
             headers={"Accept": "*/*"},
-            params={"token": "testtoken"},
             stream=True,
         )