+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The intended audience of this file is for py42 consumers -- as such, changes that don't affect
 how a consumer would use the library (e.g. adding unit tests, updating documentation, etc) are not captured here.
 
+## 1.29.0 - 2025-05-06
+
+### Fixed
+
+- An issue where file download methods did not work in some conditions.
+
+### Changed
+
+- The internal method `ExfiltratedDataService.get_download_token` now takes only one parameter, downloadRequestUrl. This is the full URL (with parameters) that will be used to request the download token.
+
 ## 1.28.0 - 2025-03-21
 
 ### Deprecated